Moved single and multi-stage helpers out to be general in preparation for extraction test refactoring.
