+++ tests/MEDS_Transforms/transform_tester_base.py
@@ -11,21 +11,16 @@ try:
 except ImportError:
     from yaml import Loader
 
-import json
 import os
-import tempfile
 from collections import defaultdict
-from contextlib import contextmanager
 from io import StringIO
 from pathlib import Path
 
-import numpy as np
 import polars as pl
 import rootutils
 from meds import subject_id_field
-from nested_ragged_tensors.ragged_numpy import JointNestedRaggedTensorDict
 
-from tests.utils import MEDS_PL_SCHEMA, assert_df_equal, parse_meds_csvs, run_command
+from tests.utils import FILE_T, MEDS_PL_SCHEMA, multi_stage_tester, parse_meds_csvs, single_stage_tester
 
 root = rootutils.setup_root(__file__, dotenv=True, pythonpath=True, cwd=True)
 
@@ -223,177 +218,45 @@ def parse_code_metadata_csv(csv_str: str) -> pl.DataFrame:
 MEDS_CODE_METADATA = parse_code_metadata_csv(MEDS_CODE_METADATA_CSV)
 
 
-def check_NRT_output(
-    output_fp: Path,
-    want_nrt: JointNestedRaggedTensorDict,
-):
-    assert output_fp.is_file(), f"Expected {output_fp} to exist."
-
-    got_nrt = JointNestedRaggedTensorDict.load(output_fp)
-
-    # assert got_nrt.schema == want_nrt.schema, (
-    #    f"Expected the schema of the NRT at {output_fp} to be equal to the target.\n"
-    #    f"Wanted:\n{want_nrt.schema}\n"
-    #    f"Got:\n{got_nrt.schema}"
-    # )
-
-    want_tensors = want_nrt.tensors
-    got_tensors = got_nrt.tensors
-
-    assert got_tensors.keys() == want_tensors.keys(), (
-        f"Expected the keys of the NRT at {output_fp} to be equal to the target.\n"
-        f"Wanted:\n{list(want_tensors.keys())}\n"
-        f"Got:\n{list(got_tensors.keys())}"
-    )
-
-    for k in want_tensors.keys():
-        want_v = want_tensors[k]
-        got_v = got_tensors[k]
-
-        assert type(want_v) is type(got_v), (
-            f"Expected tensor {k} of the NRT at {output_fp} to be of the same type as the target.\n"
-            f"Wanted:\n{type(want_v)}\n"
-            f"Got:\n{type(got_v)}"
-        )
-
-        if isinstance(want_v, list):
-            assert len(want_v) == len(got_v), (
-                f"Expected list {k} of the NRT at {output_fp} to be of the same length as the target.\n"
-                f"Wanted:\n{len(want_v)}\n"
-                f"Got:\n{len(got_v)}"
-            )
-            for i, (want_i, got_i) in enumerate(zip(want_v, got_v)):
-                assert np.array_equal(want_i, got_i, equal_nan=True), (
-                    f"Expected tensor {k}[{i}] of the NRT at {output_fp} to be equal to the target.\n"
-                    f"Wanted:\n{want_i}\n"
-                    f"Got:\n{got_i}"
-                )
-        else:
-            assert np.array_equal(want_v, got_v, equal_nan=True), (
-                f"Expected tensor {k} of the NRT at {output_fp} to be equal to the target.\n"
-                f"Wanted:\n{want_v}\n"
-                f"Got:\n{got_v}"
-            )
-
-
-def check_df_output(
-    output_fp: Path,
-    want_df: pl.DataFrame,
-    check_column_order: bool = False,
-    check_row_order: bool = True,
-    **kwargs,
-):
-    assert output_fp.is_file(), f"Expected {output_fp} to exist."
-
-    got_df = pl.read_parquet(output_fp, glob=False)
-    assert_df_equal(
-        want_df,
-        got_df,
-        (f"Expected the dataframe at {output_fp} to be equal to the target.\n"),
-        check_column_order=check_column_order,
-        check_row_order=check_row_order,
-        **kwargs,
-    )
-
-
-@contextmanager
-def input_MEDS_dataset(
+def remap_inputs_for_transform(
     input_code_metadata: pl.DataFrame | str | None = None,
     input_shards: dict[str, pl.DataFrame] | None = None,
     input_shards_map: dict[str, list[int]] | None = None,
     input_splits_map: dict[str, list[int]] | None = None,
-):
-    with tempfile.TemporaryDirectory() as d:
-        MEDS_dir = Path(d) / "MEDS_cohort"
-        cohort_dir = Path(d) / "output_cohort"
+) -> dict[str, FILE_T]:
+    unified_inputs = {}
 
-        MEDS_data_dir = MEDS_dir / "data"
-        MEDS_metadata_dir = MEDS_dir / "metadata"
+    if input_code_metadata is None:
+        input_code_metadata = MEDS_CODE_METADATA
+    elif isinstance(input_code_metadata, str):
+        input_code_metadata = parse_code_metadata_csv(input_code_metadata)
 
-        # Create the directories
-        MEDS_data_dir.mkdir(parents=True)
-        MEDS_metadata_dir.mkdir(parents=True)
-        cohort_dir.mkdir(parents=True)
+    unified_inputs["metadata/codes.parquet"] = input_code_metadata
 
-        # Write the shards map
-        if input_shards_map is None:
-            input_shards_map = SHARDS
+    if input_shards is None:
+        input_shards = MEDS_SHARDS
 
-        shards_fp = MEDS_metadata_dir / ".shards.json"
-        shards_fp.write_text(json.dumps(input_shards_map))
+    for shard_name, df in input_shards.items():
+        unified_inputs[f"data/{shard_name}.parquet"] = df
 
-        # Write the splits parquet file
-        if input_splits_map is None:
-            input_splits_map = SPLITS
-        input_splits_as_df = defaultdict(list)
-        for split_name, subject_ids in input_splits_map.items():
-            input_splits_as_df[subject_id_field].extend(subject_ids)
-            input_splits_as_df["split"].extend([split_name] * len(subject_ids))
-        input_splits_df = pl.DataFrame(input_splits_as_df)
-        input_splits_fp = MEDS_metadata_dir / "subject_splits.parquet"
-        input_splits_df.write_parquet(input_splits_fp, use_pyarrow=True)
+    if input_shards_map is None:
+        input_shards_map = SHARDS
 
-        if input_shards is None:
-            input_shards = MEDS_SHARDS
+    unified_inputs["metadata/.shards.json"] = input_shards_map
 
-        # Write the shards
-        for shard_name, df in input_shards.items():
-            fp = MEDS_data_dir / f"{shard_name}.parquet"
-            fp.parent.mkdir(parents=True, exist_ok=True)
-            df.write_parquet(fp, use_pyarrow=True)
+    if input_splits_map is None:
+        input_splits_map = SPLITS
 
-        code_metadata_fp = MEDS_metadata_dir / "codes.parquet"
-        if input_code_metadata is None:
-            input_code_metadata = MEDS_CODE_METADATA
-        elif isinstance(input_code_metadata, str):
-            input_code_metadata = parse_code_metadata_csv(input_code_metadata)
-        input_code_metadata.write_parquet(code_metadata_fp, use_pyarrow=True)
+    input_splits_as_df = defaultdict(list)
+    for split_name, subject_ids in input_splits_map.items():
+        input_splits_as_df[subject_id_field].extend(subject_ids)
+        input_splits_as_df["split"].extend([split_name] * len(subject_ids))
 
-        yield MEDS_dir, cohort_dir
+    input_splits_df = pl.DataFrame(input_splits_as_df)
 
+    unified_inputs["metadata/subject_splits.parquet"] = input_splits_df
 
-def check_outputs(
-    cohort_dir: Path,
-    want_data: dict[str, pl.DataFrame] | None = None,
-    want_metadata: dict[str, pl.DataFrame] | pl.DataFrame | None = None,
-    assert_no_other_outputs: bool = True,
-    outputs_from_cohort_dir: bool = False,
-):
-    if want_metadata is not None:
-        if isinstance(want_metadata, pl.DataFrame):
-            want_metadata = {"codes.parquet": want_metadata}
-        metadata_root = cohort_dir if outputs_from_cohort_dir else cohort_dir / "metadata"
-        for shard_name, want in want_metadata.items():
-            if Path(shard_name).suffix == "":
-                shard_name = f"{shard_name}.parquet"
-            check_df_output(metadata_root / shard_name, want)
-
-    if want_data:
-        data_root = cohort_dir if outputs_from_cohort_dir else cohort_dir / "data"
-        all_file_suffixes = set()
-        for shard_name, want in want_data.items():
-            if Path(shard_name).suffix == "":
-                shard_name = f"{shard_name}.parquet"
-
-            file_suffix = Path(shard_name).suffix
-            all_file_suffixes.add(file_suffix)
-
-            output_fp = data_root / f"{shard_name}"
-            if file_suffix == ".parquet":
-                check_df_output(output_fp, want)
-            elif file_suffix == ".nrt":
-                check_NRT_output(output_fp, want)
-            else:
-                raise ValueError(f"Unknown file suffix: {file_suffix}")
-
-        if assert_no_other_outputs:
-            all_outputs = []
-            for suffix in all_file_suffixes:
-                all_outputs.extend(list((data_root).glob(f"**/*{suffix}")))
-            assert len(want_data) == len(all_outputs), (
-                f"Want {len(want_data)} outputs, but found {len(all_outputs)}.\n"
-                f"Found outputs: {[fp.relative_to(data_root) for fp in all_outputs]}\n"
-            )
+    return unified_inputs
 
 
 def single_stage_transform_tester(
@@ -408,43 +271,28 @@ def single_stage_transform_tester(
     should_error: bool = False,
     **input_data_kwargs,
 ):
-    with input_MEDS_dataset(**input_data_kwargs) as (MEDS_dir, cohort_dir):
-        pipeline_config_kwargs = {
-            "input_dir": str(MEDS_dir.resolve()),
-            "cohort_dir": str(cohort_dir.resolve()),
-            "stages": [stage_name],
-            "hydra.verbose": True,
-        }
+    base_kwargs = {
+        "script": transform_script,
+        "stage_name": stage_name,
+        "stage_kwargs": transform_stage_kwargs,
+        "do_pass_stage_name": do_pass_stage_name,
+        "do_use_config_yaml": do_use_config_yaml,
+        "assert_no_other_outputs": assert_no_other_outputs,
+        "should_error": should_error,
+        "config_name": "preprocess",
+        "input_files": remap_inputs_for_transform(**input_data_kwargs),
+    }
 
-        if transform_stage_kwargs:
-            pipeline_config_kwargs["stage_configs"] = {stage_name: transform_stage_kwargs}
+    want_outputs = {}
+    if want_data:
+        for data_fn, want in want_data.items():
+            want_outputs[f"data/{data_fn}"] = want
+    if want_metadata is not None:
+        want_outputs["metadata/codes.parquet"] = want_metadata
 
-        run_command_kwargs = {
-            "script": transform_script,
-            "hydra_kwargs": pipeline_config_kwargs,
-            "test_name": f"Single stage transform: {stage_name}",
-            "should_error": should_error,
-        }
-        if do_use_config_yaml:
-            run_command_kwargs["do_use_config_yaml"] = True
-            run_command_kwargs["config_name"] = "preprocess"
-        if do_pass_stage_name:
-            run_command_kwargs["stage"] = stage_name
-            run_command_kwargs["do_pass_stage_name"] = True
+    base_kwargs["want_outputs"] = want_outputs
 
-        # Run the transform
-        stderr, stdout = run_command(**run_command_kwargs)
-        if should_error:
-            return
-
-        try:
-            check_outputs(cohort_dir, want_data=want_data, want_metadata=want_metadata)
-        except Exception as e:
-            raise AssertionError(
-                f"Single stage transform {stage_name} failed.\n"
-                f"Script stdout:\n{stdout}\n"
-                f"Script stderr:\n{stderr}"
-            ) from e
+    single_stage_tester(**base_kwargs)
 
 
 def multi_stage_transform_tester(
@@ -454,55 +302,17 @@ def multi_stage_transform_tester(
     do_pass_stage_name: bool | dict[str, bool] = True,
     want_data: dict[str, pl.DataFrame] | None = None,
     want_metadata: pl.DataFrame | None = None,
-    outputs_from_cohort_dir: bool = True,
     **input_data_kwargs,
 ):
-    with input_MEDS_dataset(**input_data_kwargs) as (MEDS_dir, cohort_dir):
-        match stage_configs:
-            case None:
-                stage_configs = {}
-            case str():
-                stage_configs = load_yaml(stage_configs, Loader=Loader)
-            case dict():
-                pass
-            case _:
-                raise ValueError(f"Unknown stage_configs type: {type(stage_configs)}")
+    base_kwargs = {
+        "scripts": transform_scripts,
+        "stage_names": stage_names,
+        "stage_configs": stage_configs,
+        "do_pass_stage_name": do_pass_stage_name,
+        "assert_no_other_outputs": False,  # TODO(mmd): eventually fix
+        "config_name": "preprocess",
+        "input_files": remap_inputs_for_transform(**input_data_kwargs),
+        "want_outputs": {**want_data, **want_metadata},
+    }
 
-        match do_pass_stage_name:
-            case True:
-                do_pass_stage_name = {stage_name: True for stage_name in stage_names}
-            case False:
-                do_pass_stage_name = {stage_name: False for stage_name in stage_names}
-            case dict():
-                pass
-            case _:
-                raise ValueError(f"Unknown do_pass_stage_name type: {type(do_pass_stage_name)}")
-
-        pipeline_config_kwargs = {
-            "input_dir": str(MEDS_dir.resolve()),
-            "cohort_dir": str(cohort_dir.resolve()),
-            "stages": stage_names,
-            "stage_configs": stage_configs,
-            "hydra.verbose": True,
-        }
-
-        script_outputs = {}
-        n_stages = len(stage_names)
-        for i, (stage, script) in enumerate(zip(stage_names, transform_scripts)):
-            script_outputs[stage] = run_command(
-                script=script,
-                hydra_kwargs=pipeline_config_kwargs,
-                do_use_config_yaml=True,
-                config_name="preprocess",
-                test_name=f"Multi stage transform {i}/{n_stages}: {stage}",
-                stage_name=stage,
-                do_pass_stage_name=do_pass_stage_name[stage],
-            )
-
-        check_outputs(
-            cohort_dir,
-            want_data=want_data,
-            want_metadata=want_metadata,
-            outputs_from_cohort_dir=outputs_from_cohort_dir,
-            assert_no_other_outputs=False,  # this currently doesn't work due to metadata / data confusions.
-        )
+    multi_stage_tester(**base_kwargs)
