+++ tests/utils.py
@@ -1,11 +1,22 @@
+import json
 import subprocess
 import tempfile
+from contextlib import contextmanager
 from io import StringIO
 from pathlib import Path
+from typing import Any
 
+import numpy as np
 import polars as pl
+from nested_ragged_tensors.ragged_numpy import JointNestedRaggedTensorDict
 from omegaconf import OmegaConf
 from polars.testing import assert_frame_equal
+from yaml import load as load_yaml
+
+try:
+    from yaml import CLoader as Loader
+except ImportError:
+    from yaml import Loader
 
 DEFAULT_CSV_TS_FORMAT = "%m/%d/%Y, %H:%M:%S"
 
@@ -192,3 +203,256 @@ def assert_df_equal(want: pl.DataFrame, got: pl.DataFrame, msg: str = None, **kw
         print("got:")
         print(got)
         raise AssertionError(f"{msg}\n{e}") from e
+
+
+def check_NRT_output(
+    output_fp: Path,
+    want_nrt: JointNestedRaggedTensorDict,
+):
+    assert output_fp.is_file(), f"Expected {output_fp} to exist."
+
+    got_nrt = JointNestedRaggedTensorDict.load(output_fp)
+
+    # assert got_nrt.schema == want_nrt.schema, (
+    #    f"Expected the schema of the NRT at {output_fp} to be equal to the target.\n"
+    #    f"Wanted:\n{want_nrt.schema}\n"
+    #    f"Got:\n{got_nrt.schema}"
+    # )
+
+    want_tensors = want_nrt.tensors
+    got_tensors = got_nrt.tensors
+
+    assert got_tensors.keys() == want_tensors.keys(), (
+        f"Expected the keys of the NRT at {output_fp} to be equal to the target.\n"
+        f"Wanted:\n{list(want_tensors.keys())}\n"
+        f"Got:\n{list(got_tensors.keys())}"
+    )
+
+    for k in want_tensors.keys():
+        want_v = want_tensors[k]
+        got_v = got_tensors[k]
+
+        assert type(want_v) is type(got_v), (
+            f"Expected tensor {k} of the NRT at {output_fp} to be of the same type as the target.\n"
+            f"Wanted:\n{type(want_v)}\n"
+            f"Got:\n{type(got_v)}"
+        )
+
+        if isinstance(want_v, list):
+            assert len(want_v) == len(got_v), (
+                f"Expected list {k} of the NRT at {output_fp} to be of the same length as the target.\n"
+                f"Wanted:\n{len(want_v)}\n"
+                f"Got:\n{len(got_v)}"
+            )
+            for i, (want_i, got_i) in enumerate(zip(want_v, got_v)):
+                assert np.array_equal(want_i, got_i, equal_nan=True), (
+                    f"Expected tensor {k}[{i}] of the NRT at {output_fp} to be equal to the target.\n"
+                    f"Wanted:\n{want_i}\n"
+                    f"Got:\n{got_i}"
+                )
+        else:
+            assert np.array_equal(want_v, got_v, equal_nan=True), (
+                f"Expected tensor {k} of the NRT at {output_fp} to be equal to the target.\n"
+                f"Wanted:\n{want_v}\n"
+                f"Got:\n{got_v}"
+            )
+
+
+def check_df_output(
+    output_fp: Path,
+    want_df: pl.DataFrame,
+    check_column_order: bool = False,
+    check_row_order: bool = True,
+    **kwargs,
+):
+    assert output_fp.is_file(), f"Expected {output_fp} to exist."
+
+    got_df = pl.read_parquet(output_fp, glob=False)
+    assert_df_equal(
+        want_df,
+        got_df,
+        (f"Expected the dataframe at {output_fp} to be equal to the target.\n"),
+        check_column_order=check_column_order,
+        check_row_order=check_row_order,
+        **kwargs,
+    )
+
+
+FILE_T = pl.DataFrame | dict[str, Any]
+
+
+@contextmanager
+def input_dataset(input_files: dict[str, FILE_T] | None = None):
+    with tempfile.TemporaryDirectory() as d:
+        input_dir = Path(d) / "input_cohort"
+        cohort_dir = Path(d) / "output_cohort"
+
+        for filename, data in input_files.items():
+            fp = input_dir / filename
+            fp.parent.mkdir(parents=True, exist_ok=True)
+
+            match data:
+                case pl.DataFrame() if fp.suffix == "":
+                    data.write_parquet(fp.with_suffix(".parquet"), use_pyarrow=True)
+                case pl.DataFrame() if fp.suffix == ".parquet":
+                    data.write_parquet(fp, use_pyarrow=True)
+                case dict() if fp.suffix == "":
+                    fp.with_suffix(".json").write_text(json.dumps(data))
+                case dict() if fp.suffix.endswith(".json"):
+                    fp.write_text(json.dumps(data))
+                case _:
+                    raise ValueError(f"Unknown data type {type(data)} for file {fp.relative_to(input_dir)}")
+
+        yield input_dir, cohort_dir
+
+
+def check_outputs(
+    cohort_dir: Path,
+    want_outputs: dict[str, pl.DataFrame],
+    assert_no_other_outputs: bool = True,
+):
+    all_file_suffixes = set()
+
+    for output_name, want in want_outputs.items():
+        if Path(output_name).suffix == "":
+            output_name = f"{output_name}.parquet"
+
+        file_suffix = Path(output_name).suffix
+        all_file_suffixes.add(file_suffix)
+
+        output_fp = cohort_dir / output_name
+
+        if not output_fp.is_file():
+            raise AssertionError(f"Expected {output_fp} to exist.")
+
+        match file_suffix:
+            case ".parquet":
+                check_df_output(output_fp, want)
+            case ".nrt":
+                check_NRT_output(output_fp, want)
+            case _:
+                raise ValueError(f"Unknown file suffix: {file_suffix}")
+
+    if assert_no_other_outputs:
+        all_outputs = []
+        for suffix in all_file_suffixes:
+            all_outputs.extend(list(cohort_dir.glob(f"**/*{suffix}")))
+        assert len(want_outputs) == len(all_outputs), (
+            f"Want {len(want_outputs)} outputs, but found {len(all_outputs)}.\n"
+            f"Found outputs: {[fp.relative_to(cohort_dir) for fp in all_outputs]}\n"
+        )
+
+
+def single_stage_tester(
+    script: str | Path,
+    stage_name: str,
+    stage_kwargs: dict[str, str] | None,
+    do_pass_stage_name: bool = False,
+    do_use_config_yaml: bool = False,
+    want_outputs: dict[str, pl.DataFrame] | None = None,
+    assert_no_other_outputs: bool = True,
+    should_error: bool = False,
+    config_name: str = "preprocess",
+    input_files: dict[str, FILE_T] | None = None,
+):
+    with input_dataset(input_files) as (input_dir, cohort_dir):
+        pipeline_config_kwargs = {
+            "input_dir": str(input_dir.resolve()),
+            "cohort_dir": str(cohort_dir.resolve()),
+            "stages": [stage_name],
+            "hydra.verbose": True,
+        }
+
+        if stage_kwargs:
+            pipeline_config_kwargs["stage_configs"] = {stage_name: stage_kwargs}
+
+        run_command_kwargs = {
+            "script": script,
+            "hydra_kwargs": pipeline_config_kwargs,
+            "test_name": f"Single stage transform: {stage_name}",
+            "should_error": should_error,
+            "config_name": config_name,
+        }
+        if do_use_config_yaml:
+            run_command_kwargs["do_use_config_yaml"] = True
+
+        if do_pass_stage_name:
+            run_command_kwargs["stage"] = stage_name
+            run_command_kwargs["do_pass_stage_name"] = True
+
+        # Run the transform
+        stderr, stdout = run_command(**run_command_kwargs)
+        if should_error:
+            return
+
+        try:
+            check_outputs(
+                cohort_dir, want_outputs=want_outputs, assert_no_other_outputs=assert_no_other_outputs
+            )
+        except Exception as e:
+            raise AssertionError(
+                f"Single stage transform {stage_name} failed.\n"
+                f"Script stdout:\n{stdout}\n"
+                f"Script stderr:\n{stderr}"
+            ) from e
+
+
+def multi_stage_tester(
+    scripts: list[str | Path],
+    stage_names: list[str],
+    stage_configs: dict[str, str] | str | None,
+    do_pass_stage_name: bool | dict[str, bool] = True,
+    want_outputs: dict[str, pl.DataFrame] | None = None,
+    assert_no_other_outputs: bool = False,
+    config_name: str = "preprocess",
+    input_files: dict[str, FILE_T] | None = None,
+    **pipeline_kwargs,
+):
+    with input_dataset(input_files) as (input_dir, cohort_dir):
+        match stage_configs:
+            case None:
+                stage_configs = {}
+            case str():
+                stage_configs = load_yaml(stage_configs, Loader=Loader)
+            case dict():
+                pass
+            case _:
+                raise ValueError(f"Unknown stage_configs type: {type(stage_configs)}")
+
+        match do_pass_stage_name:
+            case True:
+                do_pass_stage_name = {stage_name: True for stage_name in stage_names}
+            case False:
+                do_pass_stage_name = {stage_name: False for stage_name in stage_names}
+            case dict():
+                pass
+            case _:
+                raise ValueError(f"Unknown do_pass_stage_name type: {type(do_pass_stage_name)}")
+
+        pipeline_config_kwargs = {
+            "input_dir": str(input_dir.resolve()),
+            "cohort_dir": str(cohort_dir.resolve()),
+            "stages": stage_names,
+            "stage_configs": stage_configs,
+            "hydra.verbose": True,
+            **pipeline_kwargs,
+        }
+
+        script_outputs = {}
+        n_stages = len(stage_names)
+        for i, (stage, script) in enumerate(zip(stage_names, scripts)):
+            script_outputs[stage] = run_command(
+                script=script,
+                hydra_kwargs=pipeline_config_kwargs,
+                do_use_config_yaml=True,
+                config_name=config_name,
+                test_name=f"Multi stage transform {i}/{n_stages}: {stage}",
+                stage_name=stage,
+                do_pass_stage_name=do_pass_stage_name[stage],
+            )
+
+        check_outputs(
+            cohort_dir,
+            want_outputs=want_outputs,
+            assert_no_other_outputs=assert_no_other_outputs,
+        )
